requirements.txt
arcade==2.6.17
keyboard==0.13.5
Pillow==9.3.0
Pillow==10.0.0
pygame==2.4.0
PyOpenGL==3.1.7
PyOpenGL==3.1.7
PyQt5==5.15.9
PyQt5_sip==12.12.2
